pyproject.toml
[project]
name = "kalshi-market-trawler"
version = "0.1.0"
description = "Discovery and analysis pipeline for Kalshi prediction markets"
authors = [
    {name = "sevroauasmo", email = "126302013+sevroauasmo@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0",
    "requests>=2.28",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "click>=8.0",
    "rich>=13.0",
    "cryptography>=42.0",
    "scipy>=1.12",
    "prophet (>=1.3.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]

[project.scripts]
trawler = "trawler.cli:cli"

[tool.poetry]
packages = [{include = "trawler"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
